requirements.txt
chess==1.6.1
Pillow==8.3.1
